bump scipy version. remove upcast.

+++ src/aspire/operators/filters.py
@@ -333,8 +333,7 @@ class ArrayFilter(Filter):
         #  for values slightly outside the interpolation grid bounds.
         interpolator = RegularGridInterpolator(
             _input_pts,
-            # https://github.com/scipy/scipy/issues/17718
-            self.xfer_fn_array.astype(np.float64),
+            self.xfer_fn_array,
             method="linear",
             bounds_error=False,
             fill_value=None,
@@ -349,8 +348,8 @@ class ArrayFilter(Filter):
         # Result is 1 x np.prod(self.sz) in shape; convert to a 1-d vector
         result = np.squeeze(result, 0)
 
-        # Recast result with correct dtype
-        return result.astype(self.xfer_fn_array.dtype)
+        # Scipy's interpolator will upcast singles. Recasting.
+        return result.astype(self.xfer_fn_array.dtype, copy=False)
 
     def evaluate_grid(self, L, *args, dtype=np.float32, **kwargs):
         """
